requirements.txt
eth-brownie>=1.11.10,<2.0.0
dotmap==1.3.23
tabulate==0.8.7
ens-namehash==1.0.0
rich==9.3.0
python-dotenv==0.15.0
python-decouple==3.3
black==19.10b0